Handle exc_info in logger and persist requirement priority
